requirements.txt
azure-functions
psycopg2-binary
pydantic>=2.0.0
python-dotenv
